% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FACTOR-package.R
\docType{package}
\name{FACTOR-package}
\alias{FACTOR}
\alias{FACTOR-package}
\title{FACTOR: Package}
\description{
The function included in this package lets you explore power and effect size. When used with the {manipulate} package, you can use the sliders to set mu_0 and mu_A for a one-sample test (or, alternatively, think about these as mu_1 and mu_2 for a two-sample test), sigma, alpha, and n, and you can choose whether you are testing a one-sided hypothesis of mu_A being "greater" or "less" than mu_0 or are testing the two-sided hypothesis that mu_A != mu_0 ("two.tailed"). The graph will output Power and Effect Size.
}
\author{
\strong{Maintainer}: Sajana Aryal \email{sajanaaryal5566@gmail.com}

}
\keyword{internal}
